+++ txnbuild/CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this
 file.  This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased [v1.2.0] 
+
+* Added `MaxTrustlineLimit` which represents the maximum value for a trustline limit.
+* ChangeTrust operation with no `Limit` field set now defaults to `MaxTrustlineLimit`.
+
 ## [v1.1.0](https://github.com/stellar/go/releases/tag/horizonclient-v1.1.0) - 2019-02-02
 
 * Support for multiple signatures ([#1198](https://github.com/stellar/go/pull/1198))